[Feature] [NVM-141] 원문 청크 임베딩 생성 및 저장 흐름 구현

+++ app/models/upload_batch.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from uuid import UUID
 
-from sqlalchemy import DateTime, Integer, String
+from sqlalchemy import BigInteger, DateTime, Integer, String
 from sqlalchemy.dialects.postgresql import UUID as PostgresUUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -14,7 +14,7 @@ class UploadBatch(TimestampMixin, Base):
 
     id: Mapped[UUID] = mapped_column(PostgresUUID(as_uuid=True), primary_key=True)
     work_id: Mapped[UUID] = mapped_column(PostgresUUID(as_uuid=True))
-    member_id: Mapped[int] = mapped_column(Integer)
+    member_id: Mapped[int] = mapped_column(BigInteger)
     upload_type: Mapped[str] = mapped_column(String(40))
     source_type: Mapped[str] = mapped_column(String(20))
     status: Mapped[str] = mapped_column(String(20))
